+++ src/wifi/model/edca-txop-n.h
@@ -262,7 +262,7 @@ public:
   /**
    * Event handler when a Block ACK timeout has occurred.
    */
-  void MissedBlockAck (uint32_t nMpdus);
+  void MissedBlockAck (uint8_t nMpdus);
   void GotAddBaResponse (const MgtAddBaResponseHeader *respHdr, Mac48Address recipient);
   void GotDelBaFrame (const MgtDelBaHeader *delBaHdr, Mac48Address recipient);
   /**
